Move models to flask_discord_interactions.models
