+++ flask_discord_interactions/models/command.py
@@ -1,0 +1,19 @@
+class ApplicationCommandType:
+    "Represents the different application command types integers."
+    CHAT_INPUT = 1
+    USER = 2
+    MESSAGE = 3
+
+
+class CommandOptionType:
+    "Represents the different option type integers."
+    SUB_COMMAND = 1
+    SUB_COMMAND_GROUP = 2
+    STRING = 3
+    INTEGER = 4
+    BOOLEAN = 5
+    USER = 6
+    CHANNEL = 7
+    ROLE = 8
+    MENTIONABLE = 9
+    NUMBER = 10
