+++ flask_discord_interactions/models/message.py
@@ -1,14 +1,14 @@
-import json
-import inspect
 import dataclasses
+import inspect
 from typing import List, Union
+import json
+
+from flask_discord_interactions.models.utils import LoadableDataclass
+from flask_discord_interactions.models.component import Component
+from flask_discord_interactions.models.embed import Embed
 
 
-from flask_discord_interactions.embed import Embed
-from flask_discord_interactions.component import Component
-
-
-class ResponseType:
+class InteractionResponseType:
     "Represents the different response type integers."
     PONG = 1
     CHANNEL_MESSAGE_WITH_SOURCE = 4
@@ -18,10 +18,10 @@ class ResponseType:
 
 
 @dataclasses.dataclass
-class Response:
+class Message(LoadableDataclass):
     """
-    Represents a response to a Discord interaction (incoming webhook)
-    or a followup message (outgoing webhook).
+    Represents a Message, often a response to a Discord interaction (incoming
+    webhook) or a followup message (outgoing webhook).
 
     Attributes
     ----------
@@ -50,7 +50,7 @@ class Response:
     file
         The file to attach to the message. Only valid for outgoing webhooks.
     files
-        An array of files to attach to the message. Speficy just one of
+        An array of files to attach to the message. Specify just one of
         ``file`` or ``files``. Only valid for outgoing webhooks.
     components
         An array of :class:`.Component` objects representing message
@@ -81,7 +81,7 @@ class Response:
             self.files = [self.file]
 
         if self.ephemeral and self.files is not None:
-            raise ValueError("Ephemeral responses cannot include files.")
+            raise ValueError("Ephemeral Messages cannot include files.")
 
         if self.embeds is not None:
             for i, embed in enumerate(self.embeds):
@@ -90,26 +90,26 @@ class Response:
 
         if self.update:
             if self.deferred:
-                self.response_type = ResponseType.DEFERRED_UPDATE_MESSAGE
+                self.Message_type = InteractionResponseType.DEFERRED_UPDATE_MESSAGE
             else:
-                self.response_type = ResponseType.UPDATE_MESSAGE
+                self.Message_type = InteractionResponseType.UPDATE_MESSAGE
         else:
             if self.deferred:
-                self.response_type = \
-                    ResponseType.DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE
+                self.Message_type = \
+                    InteractionResponseType.DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE
             else:
-                self.response_type = ResponseType.CHANNEL_MESSAGE_WITH_SOURCE
+                self.Message_type = InteractionResponseType.CHANNEL_MESSAGE_WITH_SOURCE
 
     @property
     def flags(self):
         """
-        The flags sent with this response, determined by whether it is
+        The flags sent with this Message, determined by whether it is
         ephemeral.
         """
         return 64 if self.ephemeral else 0
 
     def dump_embeds(self):
-        "Returns the embeds of this Response as a list of dicts."
+        "Returns the embeds of this Message as a list of dicts."
         return [embed.dump() for embed in self.embeds] if self.embeds else None
 
     def dump_components(self):
@@ -119,14 +119,14 @@ class Response:
     @classmethod
     def from_return_value(cls, result):
         """
-        Convert a function return value into a Response object.
-        Converts ``None`` to an empty response, or any other type to ``str``
+        Convert a function return value into a Message object.
+        Converts ``None`` to an empty Message, or any other type to ``str``
         as message content.
 
         Parameters
         ----------
         result
-            The function return value to convert into a ``Response`` object.
+            The function return value to convert into a ``Message`` object.
         """
 
         async def construct_async(result):
@@ -141,20 +141,9 @@ class Response:
         else:
             return cls(str(result))
 
-    @classmethod
-    def from_message_command(cls, data):
-        """
-        Construct a Response object from an incoming Message Command
-        Interaction.
-        """
-
-        return cls(
-            content=data["content"],
-        )
-
     def dump(self):
         """
-        Return this ``Response`` as a dict to be sent in response to an
+        Return this ``Message`` as a dict to be sent in Message to an
         incoming webhook.
         """
 
@@ -165,13 +154,13 @@ class Response:
 
         if self.files:
             raise ValueError(
-                "files are not allowed in an initial Interaction response")
+                "files are not allowed in an initial Interaction Message")
 
         if self.update:
             raise ValueError("update is only valid for custom ID handlers")
 
         return {
-            "type": self.response_type,
+            "type": self.Message_type,
             "data": {
                 "content": self.content,
                 "tts": self.tts,
@@ -184,16 +173,16 @@ class Response:
 
     def dump_handler(self):
         """
-        Return this ``Response`` as a dict to be sent in reply to a Message
+        Return this ``Message`` as a dict to be sent in reply to a Message
         Component interaction.
         """
 
         if self.files:
             raise ValueError(
-                "files are not allowed in a custom handler response")
+                "files are not allowed in a custom handler Message")
 
         return {
-            "type": self.response_type,
+            "type": self.Message_type,
             "data": {
                 "content": self.content,
                 "tts": self.tts,
@@ -207,7 +196,7 @@ class Response:
 
     def dump_followup(self):
         """
-        Return this ``Response`` as a dict to be sent to an outgoing webhook.
+        Return this ``Message`` as a dict to be sent to an outgoing webhook.
         """
 
         if (self.content is None and self.embeds is None
@@ -217,7 +206,7 @@ class Response:
 
         if self.ephemeral or self.deferred or self.update:
             raise ValueError(
-                "ephemeral and deferred are not valid in followup responses")
+                "ephemeral and deferred are not valid in followup Messages")
 
         return {
             "content": self.content,
@@ -229,7 +218,7 @@ class Response:
 
     def dump_multipart(self):
         """
-        Return this ``Response`` to be sent to an outgoing webhook.
+        Return this ``Message`` to be sent to an outgoing webhook.
         Handles multipart encoding for file attachments.
 
         Returns an object that may have ``data``, ``files``, or ``json``
